re-org cyclepoints
