+++ bycycle/cyclepoints/__init__.py
@@ -1,4 +1,5 @@
 """Find & analyze cyclepoints."""
 
-from .extrema import find_extrema, find_zerox
+from .zerox import find_zerox
+from .extrema import find_extrema
 from .phase import extrema_interpolated_phase
