+++ bycycle/tests/cyclepoints/test_extrema.py
@@ -1,4 +1,4 @@
-"""Tests for cyclepoints.cyclepoints."""
+"""Tests for cyclepoints.extrema."""
 
 import os
 
@@ -46,26 +46,3 @@ def test_find_extrema(first_extrema):
         np.testing.assert_allclose(troughs[:len(peaks)], minima[0][:len(peaks)])
     elif first_extrema == 'peak':
         assert peaks[0] < troughs[0]
-
-
-def test_find_zerox():
-    """Test ability to find peaks and troughs."""
-
-    # Load signal
-    sig = np.load(DATA_PATH + 'sim_stationary.npy')
-
-    fs = 1000
-    f_range = (6, 14)
-
-    # Find peaks and troughs
-    peaks, troughs = find_extrema(sig, fs, f_range, boundary=1, first_extrema='peak')
-
-    # Find zerocrossings
-    rises, decays = find_zerox(sig, peaks, troughs)
-
-    assert len(peaks) == (len(rises) + 1)
-    assert len(troughs) == len(decays)
-    assert peaks[0] < decays[0]
-    assert decays[0] < troughs[0]
-    assert troughs[0] < rises[0]
-    assert rises[0] < peaks[1]
